reveals old mac keys

+++ src/list.c
@@ -97,10 +97,11 @@ list_element_t *list_remove_element(const list_element_t * wanted,
 	return head;
 }
 
-int list_len(list_element_t *head)
+// XXX: check
+size_t list_len(list_element_t *head)
 {
         list_element_t *cursor = head;
-        int size = 0;
+        size_t size = 0;
 
         while (cursor != NULL) {
                 ++size;
