+++ src/list.h
@@ -21,6 +21,6 @@ list_element_t *list_get_by_value(const void *wanted, list_element_t * head);
 list_element_t *list_remove_element(const list_element_t * wanted,
 				    list_element_t * head);
 
-int list_len(list_element_t * head);
+size_t list_len(list_element_t * head);
 
 #endif
